backend-example/vectorstore.py
"""
Módulo para manejar la base de datos vectorial

Puedes usar diferentes opciones:
- ChromaDB (local, fácil de empezar)
- Pinecone (cloud, escalable)
- Weaviate (cloud o self-hosted)
- FAISS (local, rápido)
"""

from typing import List
import os

# Ejemplo con ChromaDB (descomenta para usar)
"""
import chromadb
from chromadb.config import Settings

def initialize_vectorstore():
    client = chromadb.Client(Settings(
        chroma_db_impl="duckdb+parquet",
        persist_directory="./chroma_db"
    ))
    
    collection = client.get_or_create_collection(
        name="portfolio_knowledge",
        metadata={"hnsw:space": "cosine"}
    )
    
    return collection

def add_documents(collection, documents: List[dict]):
    collection.add(
        documents=[doc["content"] for doc in documents],
        metadatas=[doc["metadata"] for doc in documents],
        ids=[doc["id"] for doc in documents]
    )

def query_vectorstore(collection, query: str, n_results: int = 3):
    results = collection.query(
        query_texts=[query],
        n_results=n_results
    )
    return results
"""

# Ejemplo con Pinecone (descomenta para usar)
"""
import pinecone
from langchain_openai import OpenAIEmbeddings
from langchain_pinecone import PineconeVectorStore

def initialize_vectorstore():
    pinecone.init(
        api_key=os.getenv("PINECONE_API_KEY"),
        environment=os.getenv("PINECONE_ENVIRONMENT")
    )
    
    index_name = "portfolio-knowledge"
    embeddings = OpenAIEmbeddings(openai_api_key=os.getenv("OPENAI_API_KEY"))
    
    vectorstore = PineconeVectorStore(
        index_name=index_name,
        embedding=embeddings
    )
    
    return vectorstore
"""

# Placeholder function
def initialize_vectorstore():
    """
    Inicializa tu base de datos vectorial preferida
    Descomenta el código de arriba según tu elección
    """
    print("⚠️ Vectorstore no configurado. Por favor implementa initialize_vectorstore()")
    return None
